adjusted JobIsRunningFlow
